requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
billiard==3.6.3.0
black==19.10b0
celery==4.4.7
certifi==2020.6.20
click==7.1.2
distlib==0.3.1
Django==3.0.9
djangorestframework==3.11.0
filelock==3.0.12
flake8==3.8.3
importlib-metadata==1.7.0
kombu==4.6.11
mccabe==0.6.1
pathspec==0.8.0
pipenv==2020.6.2
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
redis==3.5.3
regex==2020.7.14
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
vine==1.3.0
virtualenv==20.0.30
virtualenv-clone==0.5.4
zipp==3.1.0
